Add molecule split comparison functions
InchIs are split into molecules and compared. This is done in two ways: 1) UniChem-style where comparisons are categorised into "whole <-> whole", "component <-> whole", "whole <-> component" and "component <-> component", and 2) which finds the best matching subset of molecules between the InchIs and returns the difference between those (as well as the InchI subsets and the proportion that each subset is of the original InchI).